gcp/requirements.txt
tensorflow==2.7.0
google-cloud-storage==1.16.1
Pillow==6.0.0